cps working on flair

+++ dl_toolbox/modules/BCE.py
@@ -62,7 +62,7 @@ class Multilabel(pl.LightningModule):
     def training_step(self, batch, batch_idx):
         batch = batch["sup"]
         inputs = batch["image"]
-        labels = batch["label"].long()
+        labels = batch["label"]
         onehot_labels = self.one_hot(labels)
         logits = self.network(inputs)  # B,C or C-1,H,W
         bce = self.bce(logits, onehot_labels)
@@ -73,7 +73,7 @@ class Multilabel(pl.LightningModule):
 
     def validation_step(self, batch, batch_idx):
         inputs = batch["image"]
-        labels = batch["label"].long()
+        labels = batch["label"]
         onehot_labels = self.one_hot(labels)
         logits = self.forward(inputs)
         bce = self.bce(logits, onehot_labels)
